+++ Engine/src/UI/Widgets/HyProgressBar.cpp
@@ -137,7 +137,7 @@ void HyProgressBar::SetNumFormat(HyNumberFormat format)
 /*virtual*/ void HyProgressBar::OnUpdate() /*override*/
 {
 	if(m_pBar)
-		m_pBar->SetScissor(0, 0, m_BarScissorAmt.Get(), m_pBar->GetSceneHeight());
+		m_pBar->SetScissor(0, 0, static_cast<uint32>(m_BarScissorAmt.Get()), static_cast<uint32>(m_pBar->GetSceneHeight()));
 }
 
 /*virtual*/ void HyProgressBar::OnSetup() /*override*/
